+++ src/MainComponent.h
@@ -5,7 +5,7 @@
 // have called `juce_generate_juce_header(<thisTarget>)` in your CMakeLists.txt,
 // you could `#include <JuceHeader.h>` here instead, to make all your module headers visible.
 #include <juce_gui_extra/juce_gui_extra.h>
-#include "app/App.h"
+#include "ui/App.h"
 #include "csound.hpp"
 #include "csPerfThread.hpp"
