+++ src/ui/Palette.cpp
