+++ src/ui/widgets/Dot.cpp
