+++ src/ui/widgets/ToggleGroup.cpp
